feat!(rmcp): hide type state from API
BREAKING CHANGE
