+++ src/lib.rs
@@ -73,6 +73,7 @@ where
             Some((a, c)) => RequestTargetAddress::BmcOrIpmb(a, c, LogicalUnit::Zero),
             None => RequestTargetAddress::Bmc(LogicalUnit::Zero),
         };
+
         let message = request.into();
         let (message_netfn, message_cmd) = (message.netfn(), message.cmd());
         let mut request = Request::new(message, target_address);
